+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+* New option `--min-length N%total`
+
 ## v3.7.0 (2026-02-08)
 
 * Edge Betweenness Centrality can be calculated (`--betweenness-centrality`)